+++ source/Graphics/Animation.cpp
@@ -1,5 +1,9 @@
 #include <Small/Graphics/Animation.hpp>
 
+#include <iostream>
+#include <sstream>
+#include <fstream>
+
 namespace sgl::Graphics
 {
     Animation::Animation() :
@@ -210,5 +214,150 @@ namespace sgl::Graphics
             states.texture = m_texture;
             target.draw(m_vertices, 4, sf::Quads, states);
         }
+    }
+
+    ///////////////////////////////////////////////////////////////
+
+    AnimationLoader::AnimationLoader()
+    {}
+
+    bool AnimationLoader::load(const std::string& configFile)
+    {
+        std::ifstream ifs(configFile);
+        toml::ParseResult pr = toml::parse(ifs);
+        ifs.close();
+
+        if (!pr.valid())
+        {
+            std::cerr << "AnimationLoader configuration error: " << pr.errorReason << std::endl;
+            return false;
+        }
+        else
+            m_config = pr.value;
+
+        if (m_config.find("data") == nullptr)
+        {
+            std::cerr << "AnimationLoader configuration needs a data section" << std::endl;
+            return false;
+        }
+
+        const toml::Value* spriteSheet = m_config.find("data.spriteSheet");
+        if (spriteSheet == nullptr || !spriteSheet->is<std::string>())
+        {
+            std::cerr << "AnimationLoader configuration: data.spriteSheet (string) missing" << std::endl;
+            return false;
+        }
+
+        const toml::Value* steps = m_config.find("data.steps");
+        if (steps == nullptr || !steps->is<std::vector<std::string>>())
+        {
+            std::cerr << "AnimationLoader configuration: data.steps (array of strings) missing" << std::endl;
+            return false;
+        }
+
+        const toml::Value* frameTime = m_config.find("data.frameTime");
+        if (frameTime == nullptr || !frameTime->is<double>())
+        {
+            std::cerr << "AnimationLoader configuration: data.frameTime (double) missing" << std::endl;
+            return false;
+        }
+
+        if (!m_texture.loadFromFile(spriteSheet->as<std::string>()))
+        {
+            std::cerr << "AnimationLoader configuration: couldn't load spriteSheet: " << spriteSheet->as<std::string>() << std::endl;
+            return false;
+        }
+
+        m_sprite.setFrameTime(sf::seconds(static_cast<float>(frameTime->as<double>())));
+        std::vector<std::string> stepsData = steps->as<std::vector<std::string>>();
+
+        for (std::size_t i=0, size=stepsData.size(); i < size; ++i)
+        {
+            const toml::Value* copy = m_config.find(stepsData[i] + ".copy");
+            if (copy == nullptr)
+            {
+                const toml::Value* framesCount = m_config.find(stepsData[i] + ".framesCount");
+                if (framesCount == nullptr || !framesCount->is<int>())
+                {
+                    std::cerr << "AnimationLoader configuration: " << stepsData[i] << ".framesCount is missing" << std::endl;
+                    return false;
+                }
+
+                m_animations[stepsData[i]] = Animation();
+                m_animations[stepsData[i]].setSpriteSheet(m_texture);
+
+                int framesCountData = framesCount->as<int>();
+                for (int frame=0; frame < framesCountData; ++frame)
+                {
+                    std::stringstream frameStr;
+                    frameStr << "frame" << frame;
+
+                    int top, left, width, height;
+
+                    toml::Value* check = m_config.find(stepsData[i] + "." + frameStr.str() + ".left");
+                    if (check == nullptr || !check->is<int>())
+                    {
+                        std::cerr << "AnimationLoader configuration: " << stepsData[i] << "." << frameStr.str() << ".left (integer) is missing" << std::endl;
+                        return false;
+                    }
+                    else
+                        left = check->as<int>();
+
+                    check = m_config.find(stepsData[i] + "." + frameStr.str() + ".top");
+                    if (check == nullptr || !check->is<int>())
+                    {
+                        std::cerr << "AnimationLoader configuration: " << stepsData[i] << "." << frameStr.str() << ".top (integer) is missing" << std::endl;
+                        return false;
+                    }
+                    else
+                        top = check->as<int>();
+
+                    check = m_config.find(stepsData[i] + "." + frameStr.str() + ".width");
+                    if (check == nullptr || !check->is<int>())
+                    {
+                        std::cerr << "AnimationLoader configuration: " << stepsData[i] << "." << frameStr.str() << ".width (integer) is missing" << std::endl;
+                        return false;
+                    }
+                    else
+                        width = check->as<int>();
+
+                    check = m_config.find(stepsData[i] + "." + frameStr.str() + ".height");
+                    if (check == nullptr || !check->is<int>())
+                    {
+                        std::cerr << "AnimationLoader configuration: " << stepsData[i] << "." << frameStr.str() << ".height (integer) is missing" << std::endl;
+                        return false;
+                    }
+                    else
+                        height = check->as<int>();
+
+                    m_animations[stepsData[i]].addFrame(sf::IntRect(left, top, width, height));
+                }
+            }
+            else
+            {
+                if (!copy->is<std::string>())
+                {
+                    std::cerr << "AnimationLoader configuration: " << stepsData[i] << ".copy should be a string" << std::endl;
+                    return false;
+                }
+                m_animations[stepsData[i]] = m_animations[copy->as<std::string>()];
+
+                const toml::Value* invertX = m_config.find(stepsData[i] + ".invertX");
+                if (invertX && invertX->is<bool>() && invertX->as<bool>())
+                    m_animations[stepsData[i]].invertX();
+            }
+        }
+
+        return true;
+    }
+
+    Animation& AnimationLoader::operator[](const std::string& key)
+    {
+        return m_animations[key];
+    }
+
+    AnimatedSprite& AnimationLoader::sprite()
+    {
+        return m_sprite;
     }
 }